Separate turbolinks logic into different asset
