+++ app/javascript/shopify_app/index.js
@@ -1,2 +1,3 @@
 require('./shopify_app')
 require('./flash_messages')
+require('./turbolinks')
